fix: remove not needed ConnectedRankingTable.tsx component, remove useRanking
