+++ src/components/containers/index.tsx
@@ -4,6 +4,5 @@ export { default as ConnectedPositionTable } from './ConnectedPositionTable/Conn
 export { default as ConnectedSwapForm } from './ConnectedSwapForm/ConnectedSwapForm';
 export { default as ConnectedBorrowPositionTable } from './ConnectedBorrowPositionTable/ConnectedBorrowPositionTable';
 export { default as ConnectedBorrowForm } from './ConnectedBorrowForm/ConnectedBorrowForm';
-export { default as ConnectedRankingTable } from './ConnectedRankingTable/ConnectedRankingTable';
 export { default as ConnectedMellowLPTable } from './ConnectedMellowLPTable/ConnectedMellowLPTable';
 export { default as ConnectedMellowLpDepositForm } from './ConnectedMellowLpDepositForm/ConnectedMellowLpDepositForm';
