+++ src/hooks/index.ts
@@ -19,7 +19,6 @@ export { default as useStateMemo } from './useStateMemo';
 export * from './useTokenApproval';
 export { default as useUniqueId } from './useUniqueId';
 export { default as useWallet } from './useWallet';
-export { default as useRanking } from './useRanking';
 export { default as useCurrentSeason } from './season/useCurrentSeason';
 export { default as usePastSeasons } from './season/usePastSeasons';
 export { default as useMellowLPVaults } from './useMellowLPVaults/useMellowLPVaults';
